+++ reader.h
@@ -3,7 +3,7 @@
 
 #include <stdio.h>
 #include <threads.h>
-#include <inttypes.h>
+#include <stddef.h>
 
 #define INPUT_QUEUE_CAPACITY (64 * 1024)
 
@@ -28,7 +28,7 @@ typedef struct Reader_context_t {
 void init_input_queue(Input_queue_t* iq);
 void free_input_queue(Input_queue_t* iq);
 int reader_thr(void* reader_context);
-int get_data(Input_queue_t* iq, char* output, intmax_t count);
+int get_data(Input_queue_t* iq, char* output, size_t count);
 int is_reading_finished(Input_queue_t* iq);
 Input_queue_t* init_reader(int fd); 
 char* input_queue_get_line(Input_queue_t* iq, int* new_line_found);
